Remove OpenCL dependency
The trainer should not be used in the main algorithm and because of
that, the OpenCL is not needed in the Windows Application.

It may be added in the future, but for now, it was wrapped inside ifdefs
and it can be enabled optionally.

Moreover, some of the files will be overriden by the Linux sources
in the future upon merge.